backend/requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
pillow==11.0.0
python-dotenv==1.0.1
sqlparse==0.5.2
tzdata==2024.2
